Сделал

+++ src/main/java/pro/sky/calculatorHw2/_2/controller/CalculatorController.java
@@ -21,22 +21,22 @@ public class CalculatorController {
     }
 
     @GetMapping("/plus")
-    public String sum(@RequestParam double num1, @RequestParam double num2) {
+    public String sum(@RequestParam int num1, @RequestParam int num2) {
         return num1 + "+" + num2 + "=" + calculatorService.plus(num1, num2);
     }
 
     @GetMapping("/minus")
-    public String minus(@RequestParam double num1, @RequestParam double num2) {
+    public String minus(@RequestParam int num1, @RequestParam int num2) {
         return num1 + "+" + num2 + "=" + calculatorService.minus(num1, num2);
     }
 
     @GetMapping("/multiply")
-    public String multiply(@RequestParam double num1, @RequestParam double num2) {
+    public String multiply(@RequestParam int num1, @RequestParam int num2) {
         return num1 + "+" + num2 + "=" + calculatorService.multiply(num1, num2);
     }
 
     @GetMapping("/divide")
-    public String divide(@RequestParam double num1, @RequestParam double num2) {
+    public String divide(@RequestParam int num1, @RequestParam int num2) {
         return num1 + "+" + num2 + "=" + calculatorService.divide(num1, num2);
     }
